Fix imports for registry and add missing
