+++ src/ark_msgs/joint_state.py
@@ -1,4 +1,8 @@
 # joint_state_pb2.py is generated from joint_state.proto
 from .joint_state_pb2 import JointState
+from ark_msgs.registry import msgs
+
+msgs.register_item(JointState)
+
 
 __all__ = ["JointState"]
